Refs #4. neutrons2events slight refactor: moved execute method to utils
